Add firewall configuration to IC OS.
We start by adding firewall functionality to `SetupOSConfig`.
`SetupOSConfig` has an `ICOSSettings` member which gains an optional
`FirewallSettings` (not serialized to disk if never configured).
This `ICOSSettings` struct member is shared by `HostOSConfig`
and `GuestOSConfig`.

This FirewallSettings is populated by the SetupOS `setup` command,
which can now optionally take a `--firewall-json-path` flag pointing
to a file with a list of `FirewallRule`.  If the file is not specified,
a default path from the USB setup drive config partition is tried,
and if it is absent, no firewall settings are saved.

Since we need to support legacy functionality while the config revamp
is ongoing, the setup process will copy any present firewall rules file
alongside `config.ini` to HostOS.  This file will, in turn, be injected
by `build-bootstrap-config-image.sh` into the GuestOS.  Later on, when
the second phase of the config revamp is done, and therefore migration
of IC OS settings is complete, the firewall rules files present in HostOS
and GuestOS will no longer be used because the data inside `ICOSSettings`
will be used instead.