refactor(nervous_system): Add a `Request` trait to simplify interacting with our canisters (#1091)
## Problem

Making requests to canisters in rust code is quite tedious because we
must get the request type, encode it, send it, check if the canister
returned an error, then decode the response. It is also

## Solution

Add a `Request` trait which can be used to contain the information
needed to perform a request: its method string, whether it's an update
call, and its response type.

Check out `interactions/sns_wasm.rs` to see how much it cleans up the
code.

## Notes

In the future I would like to automatically generate Request
implementations from our candid files or from the code that causes them
to be generated, but for now they are not that bad to implement
manually.

Right now, I only actually implement the trait for SNS-W's api, and only
use it in the `nervous_system/interaction` crate. I want to implement
this trait for other canisters' types as well, and use them in canister
code.

[← Previous PR](https://github.com/dfinity/ic/pull/1087)