test(nns): Support caching of the golden NNS state when running the `upgrade_canisters_with_golden_nns_state` test (#1179)
# Why

Most of the time is spent on downloading and decompressing the state.
Caching it will make debugging faster.

# What

When `USE_CACHED_STATE_DIR` is provided as an environment variable, the
test will first check whether a `nns_state` directory already exists,
and use it instead of downloading and decompressing again. If the
directory does not exist, then download and decompress to the cached
state directory. Note that there is no mechanism for cleaning the cache
within the test itself, but can be done by simply cleaning the directoy.