chore: remove uni streams from transport (#1201)
I am fairly convinced now that we can remove the uni stream support.

The value proposition of the unidirectional communication was that we
have infallible handlers so let's use unidirectional streams and not
propagate any errors to the client.

However, we are implementing a general purpose transport library that
has routing.
The routing is specified by the client's and not the framework, it does
make sense to propagate at least errors w.r.t. to routing even if the
handlers are infallible.

Another data point gRPC doesn't have anything similar to the push method
that we have. Always some application level information is passed back
from the receiver.