fix(crypto): CRP-2573 use rustls' std feature (#926)
Currently `cargo test -p ic-crypto` fails. This is because the `std`
feature of `rustls` is required but not enabled for Cargo. The Bazel
build succeeds even without the `std` feature being explicitly set for
`rustls`: this is because the `quinn` dependency implicitly activates
rustls'` `std` feature and Bazel's build semantics are different in
feature unification is always done for all crates, not just the ones
that are being compiled.

For Cargo, the `std` feature was already activated in a previous PR
(https://github.com/dfinity/ic/pull/955). This PR here ensures that both
Cargo and Bazel are in sync.

Fixes CRP-2573.