feat(PocketIC): drop ready file in PocketIC server (#1228)
This PR drops the ready file from the PocketIC server.

Originally, the ready file was added to the PocketIC server for the sake
of synchronization (or locking) - we want to have only one test in a
cargo test suite start the PocketIC server and thus the individual tests
need to agree who starts that server and what the port is:
- the port file is written atomically by the first test that manages to
write that file atomically and that test starts the server and writes
the port to the port file and afterwards creates the ready file;
- the other tests then poll for the ready file to exist and once it
exists, they all know that they can read the port from the port file.

If the other tests attempted to read from the port file without knowing
that it has been fully written, then we could get data races. However,
we can simplify the code in this PR by getting rid of the ready file by
checking for an end marker, e.g. the newline character `\n`, to learn
that the port has been fully written.