chore(nns): NeuronsFundAuditInfo migration (#1208)
## Problem

We have switched from the `hotkey_principal` field to a field named
`nns_neuron_controller` in the type `NeuronsFundNeuronPortion`. This is
because the type `nns_neuron_controller` more clearly explains the
semantics of the type than `hotkey_principal`. However, already-existing
`ProposalData`s indirectly contain `NeuronsFundNeuronPortion`s which
have `None` for the new field.

## Solution

I've implemented a migration for the `NeuronsFundNeuronPortion`s that
are transitively stored in `ProposalData` to use this new field. The
migration lives in `pre_upgrade` and iterates over the stored
`ProposalData`s, and sets their `neurons_fund_data` field to a value
that contains the migrated state.

< [Previous PR](https://github.com/dfinity/ic/pull/1171) |