fix: use only one version of ic-certification (#1267)
The external ic-certification package was used by multiple dependencies,
but with slightly different versions. This PR makes sure only one
version is used. To help solve the problem, the ic-http-proxy dependency
is also pinned to a more recent version.

Note that the external ic-certification is a different package than the
one internal in rs/certification subdir.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>