fix(execution): Canister snapshots should include Wasm globals (#1250)
Currently, the canister snapshots do not cover the state of Wasm global
variables, such that in realistic canister programs with mutable global
variables, the state is not properly recovered when a snapshot is
loaded.

This fix adds the Wasm globals to the snapshot state, similar to the
execution state. (Side note: All Wasm variables are exported by the Wasm
instrumentation.)