chore(EXC-1721): Backtrace Test Canister
Add a `rust_canister` with endpoints which crash in different ways where
we'd like to report a canister backtrace. This canister can then be used
in tests to assert that the backtrace is present. For now we just add
tests that assert the endpoint is an error.

We also need to modify the `ic-wasm` commands for the `rust_canister`
bazel rules to ensure that the name section is present in the generated
Wasm as it will be needed to produce backtraces.