chore: update query stats log message to be less confusing in PocketIC (#1322)
The log message "Current stats are uninitalized. This warning should go
away after some minutes" appears often in PocketIC logs. After some
debugging, it turned out that this is because the PocketIC instance is
not processing any query calls, but the log message indicates that it
should just go away without hinting at what needs to happen. Hence, this
PR extends the log message to be less confusing in PocketIC (and also
make sense in production).