feat(nns): Added is-commit-released.sh to nns-tools. (#1278)
Simply pass it

1. canister name
2. (query) commit ID

It then asks the canister what commit it's on, and then asks git whether
the query commit is an ancestor of the commit the canister is running.

# Future Work

## SNS

This only works for NNS. In the case of SNS, this could answer a couple
variations of the same request:

1. If an SNS had the most recent available version, would the query
commit be included?
2. Which SNSs do not yet have the query commit?

## No Canister Name

Perhaps, we can (in a future iteration), we can make this even easier by
making canister name optional. If omitted, this script could scan all
NNS canisters, and print the status of each.

## Pull Request Number Instead of Commit ID

I would also like it if you could pass a PR number, and have it figure
out the merge commit. One way to do this:

```
git log --grep ' (#1234)$' --pretty=format:"%H" master
```

This prints out the IDs of matching commits. The problem with this is
that it assumes that " (#1234)" (where 1234 is the PR number) occurs at
the end of a line. This can cause false matches. Or if the format
changes, it can have the opposite problem. Basically, this is not super
reliable, but might be good enough for the purposes of this script.