feat: guest-os qualification system test (#1227)
For long time, dre did qualification of guest-os after the rc (release
candidate) is cut successfully. We would have to wait for a successful
rc and then try and qualify a potential guest-os version (which takes
around an hour but can be flaky). Also we used staging meaning we could
only qualify one rc at a time. With this PR we want to transition to
running qualification of guest-os as part of system test suite. This PR
only adds the test itself, in the following PR's will the test be
automatically called from the CI on Daily schedule.

Qualification is run in steps:
1. Ensure the starting guest-os version is blessed
2. Upgrade the app subnets to that version
3. Upgrade system subnet to that version
4. Upgrade unassigned nodes to that version
5. Run workload tests
6. Run xnet tests
7. (Optional) - retire the initial guest-os version, and re-bless it.
Needed if the initial version contains `disk-img` instead of
`upgrade-img`
8. Downgrade app subnets to initial version
9. Downgrade system subnet to initial version
10. Downgrade unassigned nodes to initial version
11. Run workload tests
12. Run xnet tests

With this PR we should cut down on time needed to go from a green rc
towards a proposal for blessing that version and allow teams to catch
issues with qualification earlier which would add to increased
probability of release-per-week goal that we have set.

To avoid code duplication, I took the liberty of moving some code around
which is why there is so many changes but most of them are copy+paste.
Logically all the tests remain the same.

Example of successful guest os qualification run
[here](https://dash.zh1-idx1.dfinity.network/invocation/41cddade-d62d-48c2-99cd-17fd276dfc39)

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>