chore(sns): Duplicate ClaimSwapNeuronsRequest into Swap, so the version in SNS Governance can be changed independently (#1295)
This will allow us to remove the `NeuronParameters` field from
Governance. This extra rigamarole is required because the field is not
optional, so we need to make sure the callee is not expecting it before
the caller stops providing it in the request.

Once #1297 is released, this PR should be reverted. With that in mind,
I've implemented it optimizing for revert-ability. This means avoiding
doing refactors that would potentially be painful to revert, if the code
continued to change between now and then.

[Next PR →](https://github.com/dfinity/ic/pull/1297)