feat(ICRC-Ledger): add ability to read from memory manager in post_upgrade (#746)
This version of the ledger is able to read both plain stable memory and
memory manager manager stable memory in the `post_upgrade`. In order to
test the ability to read from memory manager I added the
`upgrade-to-memory-manager` feature flag. Without the flag the upgrade
writes to plain stable memory. With the flag the upgrade writes to
`UPGRADES_MEMORY` variable managed by the memory manager. In a future
commit, we will remove the flag and leave only the code that writes to
the memory manager.
This PR addresses the following tickets: FI-1409, FI-1410, FI-1411,
FI-1412 and FI-1443

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>