fix: caller of canister http composite query transform (#1337)
The caller of the canister http transform function is the management
canister when invoked as a query, but it is the anonymous principal when
invoked as a composite query. The specification prescribes the caller to
be the management canister and thus this PR fixes the caller to be the
management canister in the composite query case.