refactor(IDX): update main.sh logic (#1324)
This PR tries to simplify some of the main.sh and diff.sh logic. Some of
the big changes are:
- instead of having logic defined in both workflows.yaml and main.sh
move all the logic to shell scripts
- we don't need to check if a branch is a `hotfix` branch because all
hotfix branches are protected by default

For reviewing this PR, I would recommend opening the new files and
reading through them instead of looking at the diff. That way you can
assess if the new logic makes sense.

---------

Co-authored-by: IDX GitHub Automation <IDX GitHub Automation>
Co-authored-by: Bas van Dijk <bas@dfinity.org>