test(nns): Updated mainnet-canisters.bzl. (#1280)
That is, update our bazel configuration to reflect what is running in
production, because we just finished a round of NNS canister upgrades.

Apparently, we did not do this last time, because other canisters
(besides the ones that we upgraded today) in this file are also updated.

The changes were performed in our standard way:

```
bazel run //rs/nervous_system/tools/sync-with-released-nevous-system-wasms --config=local
```

(TBH, I don't think `--config=local` is needed, but if it is, I do not
know why.)

---------

Co-authored-by: Arshavir Ter-Gabrielyan <arshavir.ter.gabrielyan@dfinity.org>
Co-authored-by: Jason Zhu <jason.zhu@dfinity.org>