refactor(nns): Eliminate `ExecutedCreateServiceNervousSystemProposal` (#1366)
This PR refactors the code in NNS Governance responsible for executing
`CreateServiceNervousSystem` proposals s.t. it no longer requires the
internal type `ExecutedCreateServiceNervousSystemProposal`. This type
was intended to be internal but has become public for CLI tools to
validate instances of this proposal type, which adds an undesirable
crate-level dependency between, e.g., `//rs/nns/governance` and
`//rs/sns/init`.