feat(sns): Postpone enforcing the 100K SNS neuron limit till swap begins (#1332)
This PR lifts an overly strict requirement due to which some SNS
configurations were deemed invalid if it was possible to exceed the 100K
SNS neurons limit for swap participants in the worst case.

[Recently](https://github.com/dfinity/ic/pull/924), the maximum number
of swap participants has started to be enforced during swap
participation, so checking the condition upfront is no longer needed for
security, and so we lift it to allow for more SNS configurations.