feat: load number of pages from disk in merge strategy (#1248)
In order to create merge strategy, we need to know for each `PageMap`
its size on disk and in memory. We used to propagate the in memory size
from DSM thread. It is complicated code wise and makes the lazy
`PageMaps` loading harder.
This change loads pagemaps in the `TipThread` in parallel to get the
in-memory size.