refactor: Drop `CanisterQueueItem` enum (#1379)
We no longer want to enqueue multiple types of references in
`CanisterQueue`. Instead, it is more efficient to have one type of
reference and simply look it up in multiple locations (pool, map of shed
responses, map of time out responses) when appropriate.

Simplify the code by getting rid of the `CanisterQueueItem` enum.