feat(ckerc20): Add already installed canisters to ledger suite orchestrator (#1312)
XC-189: Add the ability to register via upgrade arguments ledger suites
that were initially not installed by the ledger suite orchestrator to be
able to be managed by the orchestrator, e.g. to top-up cycles or to
upgrade the underlying canisters. This is required to be able to add,
via a future NNS proposal, the ckETH ledger suite under control of the
ledger suite orchestrator.

Since the design of the orchestrator currently assumes that all managed
ledger suites corresponds to ckERC20 tokens, which would not be the case
if the ckETH ledger suite was to be managed by the orchestrator, an
important refactoring was necessary:

1. Clear separation between ckERC20 ledger suites and the other ones. In
particular, `OrchestratorInfo::managed_canisters` contains only ckERC20
ledger suites managed by the orchestrator, while the other ones are
visible via a new dedicated field
`OrchestratorInfo::managed_other_canisters`.
2. Under the hood, to avoid a migration of stable memory, a new field
was added to `ManagedCanisters` to contain the managed ledger suites
which are not ckERC20 tokens. Those two fields are kept as an
implementation detail, by exposing a `TokenId` that wraps both notions
(a ledger suite for ckERC20 token or a ledger suite for some thing
else), such that managed canisters are only accessed by `TokenId`.
3. All current tasks (top-up, archive discovery, upgrades) of the
orchestrator should work as is for all managed canisters (ckERC20 ledger
suites and the other ones).

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>