feat(IC-1579): pull out canister-based implementations of Ledger/CMC (#1386)
Pull out the canister-based implementations of some common nervous
system traits (Ledger/CMC) into a separate Bazel target. This makes it
easier to add conditionally compiled code to these implementations
(which we'll do for IC-1579, to compile the TLA instrumentation only
when `tla` and `test` features are enabled), since this minimizes the
number of new `--test_feature` Bazel targets required.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>