fix(PocketIC): drop PocketIC library timeout (#1439)
This PR drops the PocketIC library timeout of 10s when starting the
PocketIC server. Given that other parts of a test might be slow, too, it
is not clear why we'd impose a timeout on starting the PocketIC server.
Moreover, we have seen this timeout causing flakiness in tests
regularly.