feat(PocketIC): dashboard endpoint in PocketIC HTTP gateway (#1381)
This PR adds a new endpoint `/_/dashboard` of the PocketIC HTTP gateway
returning the dashboard of the underlying PocketIC instance or replica.

This way, the URL of the PocketIC HTTP gateway can be used to access the
dashboard and no separate URL of the PocketIC instance (having a
non-empty URL path of the form `/instances/<instance_id>`) or replica is
needed.

On the other hand, the path `/_/dashboard` becomes reserved and cannot
be used for frontend canister assets (note that paths with prefixes
`/api/v2` and `/api/v3` are already reserved and thus this PR does not
make a qualitative difference here).