fix(PocketIC): PocketIC server restart when using PocketIC library (#1284)
This PR makes the PocketIC library not create a file that was not
removed after the PocketIC server's TTL elapsed causing the PocketIC
library to fail when starting a new PocketIC server after the original
PocketIC server's TTL elapsed.

The motivation for adding that file was to avoid starting the PocketIC
server process at the beginning of every test in a test suite. This
reason is not particularly strong though given that running a canister
results in several processes being spawned anyway (for the sake of
canister sandboxing). Note that this PR still allows reusing the same
PocketIC server process in a test suite via the `--pid` CLI option (the
synchronization proceeds through an atomically created file by the
PocketIC server itself, with servers terminating quickly if they can't
atomically create the shared file).