Render HostOS and GuestOS firewall rules in `(hostos|guestos)_tool`.
In case of absence of the default rules file, an empty text is rendered
(empty text is valid input for the `nft` command).  If the rules file is
explicitly specified, and it does not exist, the subcommand exits with
an error.

When rules are in fact generated, in case the nftables chains they depend
upon are not defined or referenced anywhere, the chains are created prior
to flushing.  This prevents errors of coordination where perhaps the system
rules (which create the chains) have not yet been added to the system.

Both commands default to their respective OSes' config partitions to search
for the file in question (`firewall.json`).  This file is added by the
SetupOS into HostOS, and later propagated by HostOS into GuestOS when
the GuestOS is bootstrapped.

This commit does not yet incorporate the mechanisms that will actually load
the rules into the kernel.  That part is still being developed.