feat: Reduce slice limit for install_code messages in system subnets (#1401)
Reducing the slice limit for `install_code` messages is now possible
after the [update
](https://dashboard.internetcomputer.org/proposal/131699)to the DKG
interval on the NNS subnet. Given the 1T limit for `install_code`
messages on system subnets, a slice limit of 5B would mean at most 200
slices in the worst case, which is acceptable given the current DKG
interval of 500 rounds.

This change brings even closer the configuration for normal messages
(using a 2B slice limit) and `install_code` messages.