chore(ic-admin): Keep `--use-explicit-action-type` in ic-admin (#1377)
Some users already use the `--use-explicit-action-type` when using
ic-admin to create canister upgrade proposals. Avoid breaking them for
now.

The option isn't technically used. Ideally we want to have a boolean
option default to true but can be set to false, but there seems to be no
convenient way to do so with clap.