feat: [MR-523] Prevent enqueueing multiple responses for the same callback (#1071)
[[MR-523]] Currently, when inducting a response we only check that it
matches an existing callback, dropping any mismatched responses and
raising a critical error. However, a malicious subnet (or a bug) could
produce a second response for the same callback before the first one has
been executed (and the callback closed). This second response would also
get enqueued and consume an extra slot reservation; which would later
result in dropping a legitimate response due to the queue no longer
having a response slot reservation for it.

We maintain an explicit, transient set of callback IDs whose responses
are already enqueued in input queues; and check against this set for
potential duplicate responses. If a duplicate is detected, we return an
error if this is a guaranteed response (which results in a critical
error); but silently ignore duplicate best-effort responses (as this may
be a response delivered after we have already produced a local timeout
reject response).

This functionality is required in order to safely generate local timeout
reject responses for expired best-effort callbacks.

[MR-523]:
https://dfinity.atlassian.net/browse/MR-523?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ