candid(evm-rpc)!: remove `RpcService::Chain` variant (#1418)
This PR removes the `RpcService::Chain` variant (reflected in the EVM
RPC Candid interface) due to ambiguity about which RPC provider is used
to make the request.

Another option could be to always use a specific service such as [Public
Node](https://www.publicnode.com/) for these RPC requests, although this
would remain ambiguous and is only valid for some inputs compared to
specifying, for instance, `#EthMainnet(#PublicNode)`. The current
behavior is to use the RPC provider with the largest `provider_id`
matching the provided `chain_id`.

Merges into the `evm-rpc-canister` branch.