fix(cmc): Fix the data certification for the `get_average_icp_xdr_conversion_rate` endpoint (#1423)
# Why

The certification of the `get_average_icp_xdr_conversion_rate` endpoint
is wrong because it is using the wrong label. We cannot deprecate this
endpoint because the NNS Governance is using it.

# What

* The first commit adds state machine test cases for the 2 certified
endpoints (and the one for `get_average_icp_xdr_conversion_rate` fails)
* Fix the `get_average_icp_xdr_conversion_rate` endpoint in the second
commit
* Use the same version for response verification crates in the
`Cargo.toml` (as in `external_crates.bzl`)