fix(quic-transport): Send `RESET_STREAM` frame for transport handles that are dropped (#1346)
This PR adds a drop guard to wrap `quinn::SendStream` in our transport
implementation when making RPC calls. The wrapper will send a
`SendStream::reset` frame if it is dropped, which can happen if a
transport client, such as P2P, cancels the RPC. Quinn sends a `finish`
frame by default when dropping a stream, which can lead to the peer
receiving the stream thinking a complete message was sent, and attempts
to decode an incomplete message. This guard is used to send a reset
frame instead, to signal to the peer that the transmission of the
message was cancelled.

---------

Co-authored-by: tim gretler <tim.gretler@dfinity.org>