fix: adding backup if there are no active rollouts from the rollout dashboard (#1463)
This PR adds a backup option in case there are no viable subnet Git
revisions or the rollout dashboard is down. Before this PR the job would
fail and say that the matrix was invalid which results in a failed
release cut as we can see
[here](https://dfinity.slack.com/archives/C024WJPPX5W/p1726110654204759).

There are some occasions when we expect this situation with the current
code base:

1. There was no successful cuts for at least a week.
2. We didn't release because of the holidays (Like we do for Christmas) 
3. We had an urgent fix that was rolled out in a day and no new rollouts
were dispatched when the release-testing flow was issued.
4. Rollout dashboard is malfunctioning.

Yet another set of concerns need resolving. The prior version selection
algorithm (in the case of using the rollout dashboard) included all the
versions of all the ongoing rollouts. This has the following problems:

1. If there is no ongoing rollout, we obtain no version data.
2. If there are failed rollouts, but some subnets have in fact been
rolled out successfully, we excluded those versions, even though quite
logically those versions will be in mainnet if the failed rollout is the
most recent rollout.
3. If there is a hotfix rollout, which only targets a handful of
subnets, then versions may not be complete.

This PR proposes a way to only have a backup if everything is fine with
our primary source of information (rollout dashboard) but there are no
active rollouts **OR** if the rollout dashboard is unavailable. If its
available but the data is malformed the CI will panic and will require
investigation.

In addition to that we refine the algorithm used to select the versions
from the primary source of information. What we do now is:

* Go through time, from past to future, for each subnet, collecting each
version that has been targeted to that
subnet, along with its date (completion date if available, else start
time if available, else planned start time) and state.
* If the state of that subnet's rollout instance is failed, then we
disregard it. Else, we take it into account.
* Having iterated through the rollouts one by one, we eventually end up
with a list of subnets and a chronological list of
revisions that have been targeted to those subnets (and either were
successfully rolled out, or are pending and may be rolled out in the
future).
* For each subnet, we pick (based on the date we collected) the most
recent version.

Those are the versions that we're going to test against.

If this fails, that means that the versions on subnets of mainnet shall
be the initial versions for qualification.

---------

Co-authored-by: IDX GitHub Automation <IDX GitHub Automation>
Co-authored-by: Manuel Amador <manuel.amador@dfinity.org>
Co-authored-by: DFINITYManu <123385598+DFINITYManu@users.noreply.github.com>
Co-authored-by: Saša Tomić <sasa.tomic@dfinity.org>