feat: [CON-1371, CON-1372] implement the logic of stripping and assembling of block proposals (#1325)
### Description
This PR implements the stripping/assembling of the blocks.

1. We explicitly strip only the `Data` blocks since `Summary` blocks
don't contain any ingress messages
2. The newly introduced `StrippedBlockProposal` structure takes
advantage of the property of protobuf's that all fields in message are
optional: during stripping of the `BlockProposal` we thus first convert
the block to a proto and then simply set the `ingress` field to `None`.
To be able to reconstruct the block, `StrippedBlockProposal` also has a
list of ingress message ids which were stripped from the block. To
reconstruct the block we first insert the missing ingress messages into
the *proto* and only then deserialize the proto to the `BlockProposal`
type.

### Tests
Temporarily enabled stripping of the blocks in `setup_ic_network` and
ran the `consensus_performance_test` (observed 4MiB/s throughput)

### Note
This code branch is not executed yet

---------

Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>