chore: add metric for executed canisters per round (#1485)
This PR introduces a scheduler metric to count the number of canisters
actually executed in the past round.

This differs from the `scheduler_executable_canisters_per_round` metric,
which shows the number of canisters that are eligible for execution (but
may not be executed due to constraints such as round limits).