chore(sns): Remove `neuron_parameters` from ClaimSwapNeuronRequest in Swap (#1417)
This can be done safely because SNS Governance is already not expecting
this field. The field is deprecated and no longer needed.

The way that `neuron_parameters` is removed is slightly subtle.
Previously, the `ClaimSwapNeuronRequest` was duplicated between Swap and
SNS Governance. Then `neuron_parameters` was removed from the SNS
governance side. Now in this PR, they are deduplicated again (favoring
the SNS Governance side).