feat(icp-ledger): add bazel target for local replica testing (#1487)
Adding a simple bazel target to make it easier building and deploying
the icp ledger to a local ic replica using a single command.

The script run by the target also does any necessary cleanup required
between deploying a new version of the canister when an older one is
currently running there.