feat(sns): Use timer instead of heartbeat to poll archives (#1477)
# Why

In SNS Root, the polling for archives is only needed once every day, but
it is executed in every heartbeat which costs 10M instruction per round
even though it doesn't do anything.

# What

There are 4 commits:

1. Extend a current test to make sure archives can be found by SNS Root
(and test passes)
2. Remove heartbeat from SNS Root (and verify that the test fails)
3. Add timer to SNS Root for polling
4. Clean up the code only useful for heartbeat, including the
`latest_ledger_archive_poll_timestamp_seconds` in the canister state.

## Consideration for `latest_ledger_archive_poll_timestamp_seconds`

The state is protobuf, so encoding/decoding during upgrades should be
OK. If we want to rollback for any reason, it is also ok to lose the
value of `latest_ledger_archive_poll_timestamp_seconds`, so that the
first heartbeat after the rollback should start polling the archive.