refactor(nervous-system): Make ic-nervous-system-agent generic over how it calls canisters (#1495)
## Problem

The ic-nervous-system-agent library is hardcoded to only work with
ic_agent. However, we may want to use its functions in a context where
we don't have ic_agent.

## Solution

Create a `CallCanisters` trait that represents an object with the
capability to call a canister. Implement that trait for `Agent`. In the
future, implementations can also be created for other types. This is
easy to do now, since rust now finally supports async traits without
hacks.

The trait has a `call` function, and different implementations support
different errors. This is useful because different ways of calling
canisters have different plausible errors (e.g. you are not going to get
a boundary node error when a canister calls another, but you might when
an external user calls a canister).

This also means that ic-nervous-system-agent functions for calling
canisters no longer need to return `anyhow::Result`! They can instead
return the appropriate error type based on the caller's context. This is
more appropriate for libraries that want to give as much information to
their callers as possible (as opposed to `anyhow::Result` which is
pretty opaque)