feat(sns): Only set the `wasm_memory_limit` for SNS Governance when deploying an SNS (#1427)
The wasm_memory_limit setting is useful because, when you canister
reaches the limit, it prevents update calls but allows pre_upgrade to
run. This can be used to make our canisters more resilient to a
situation where they run out of memory, because it is unlikely that more
memory will be consumed in the absence of update calls, but they can
still be upgraded and keep their state because pre_upgrade is still able
to runs.

Governance does not need to be changed because we get no advantage from
having governance be at 3.5GB at this time.