feat(ckbtc): implement check_transaction of the new KYT canister [override-didc-check] (#1400)
feat: implement check_transaction interface (XC-153)

The check_transaction method checks the input addresses of a transaction
to see they can pass KYT (i.e. not in the OFAC list).

To do this, we need to fetch the actual transaction from bitcoin RPC
provider(s), as well as all its input transactions in order to compute
the input addresses. So it may take a while to return, and in certain
cases `HighLoad` or `Pending` is returned to ask the caller to call back
again later. Each call should attach at least 40 billion cycles and 0.1
billion cycles are always charged as a service fee. Unspent cycles will
be refunded.

---------

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>
Co-authored-by: gregorydemay <112856886+gregorydemay@users.noreply.github.com>