fix(cketh/ckbtc): Update fee for tECDSA signatures (#1545)
([XC-201](https://dfinity.atlassian.net/browse/XC-201)): Since the
number of nodes in the fiduciary subnet increased, the previous amount
of cycles (`25_000_000_000`) that was hard-coded in both the ckBTC and
ckETH minters to pay for tECDSA signature is no longer accurate and
signing requests will be rejected once the subnet uses 34 nodes.

The implementation switches to using `ic_cdk`, which already uses the
correct amount of cycles.

[XC-201]:
https://dfinity.atlassian.net/browse/XC-201?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ