chore: Remove testonly dependencies of ic-replay (#1548)
https://github.com/dfinity/ic/commit/44025f8cf721c87e81730001c05377c660a2d8f6
removed the synchronous QueryHandler of ic-replay in favor of the
asynchronous QueryExecutionService, which evaluates the query on the
latest certified state. During NNS recoveries, we apply extra ingress
messages to the state and expect to be able to read the resulting state.
Therefore, the change above additionally introduced "dummy
certifications" for these states created by "extra batches".

Until now, these dummy certifications were create using a function from
a test-only crate. Instead, we can simply create them with empty
signatures. This allows us to remove the test-only dependency of
ic-replay, and all of its dependants.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>