feat(NODE-1475): Configuration revamp (define config structure and config tool) (#1539)
[NODE-1360](https://dfinity.atlassian.net/browse/NODE-1360)

Changes pulled out of: https://github.com/dfinity/ic/pull/1190

This PR defines the config structure and config tool, but does not yet
integrate the config tool into SetupOS. This PR should not have an
impact on the ic-os. Later, the config tool will be utilized for config
sanitization, organization, access, and propagation.

[NODE-1360]:
https://dfinity.atlassian.net/browse/NODE-1360?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ