test(ICRC_ledger): FI-1492: Set subnet configs in golden state tests to cover all canister IDs (#1530)
Expand the canister ID ranges for the `StateMachine` app subnet in the
golden state tests for the NNS, SNS, and fiduciary subnets to cover all
possible canister IDs. This is to avoid scenarios where the tests may
fail in case there are some messages to canisters on other subnets that
cannot be routed.