feat(cmc): get default subnets query (#892)
In this PR we are looking to allow users to query the default subnets
from the cmc. Right now this information is not publically visible. In
order to be able to display the diff in proposals (when we submit
`set-authorized-subnetworks` proposals like [this
one](https://dashboard.internetcomputer.org/proposal/132409)) we've
created a simple query method that returns a vector of principals.

---------

Co-authored-by: NikolaMilosa <nikolamilosa20@gmail.com>