feat: Library that determines the initial token supply (#1494)
It does this by sending ledger a bunch of queries.

The basic idea is quite simple. Unfortunately, in practice it's more
complicated. In its basic form, the algorithm goes like this:

1. Scan transactions starting from the beginning.
2. Once we hit a transaction whose timestamp is > than the timestamp of
the first transaction, stop.
3. While scanning, notice that all transactions are minting
transactions, and total their amounts.

In practice it's slightly more difficult, because ledger can redirect
transaction queries to archive. ThickLedgerClient takes care of this.

# Future Work/Motivation

Because Dragginz did a "split" of their token, our method of assigning a
valuation to their treasury is all messed up. As a result, they are
severely restricted in the amount of SNS tokens they can take out of
their treasury, much more than we intended.

This will help us determine the inflation factor that should be applied
when assigning a valuation to their treasury. This inflation factor will
greatly reduce the valuation of their treasury, and thus, reduce the
restriction on proposals that transfer SNS tokens from the treasury.